+++ EduVPN/Networking/ServerAPIv3Handler.swift
@@ -50,7 +50,8 @@ extension ServerAPIv3Error: AppError {
             """
         case .errorGettingProfileConfig(let profile, let serverError):
             return """
-            Requested profile: \(profile.displayName.stringForCurrentLanguage())
+            Requested profile name: \(profile.displayName.stringForCurrentLanguage())
+            Requested profile id: \(profile.profileId)
             Server error: \(serverError)
             """
         case .expiresResponseHeaderIsInvalid(let value):