+++ OpenVPNTunnelExtension/PacketTunnelProvider.swift
@@ -5,6 +5,7 @@
 
 import TunnelKit
 import NetworkExtension
+import SwiftyBeaver
 
 #if os(macOS)
 enum PacketTunnelProviderError: Error {
@@ -26,9 +27,12 @@ class PacketTunnelProvider: OpenVPNTunnelProvider {
             }
         }
     }
+    #endif
 
     override func startTunnel(options: [String: NSObject]? = nil, completionHandler: @escaping (Error?) -> Void) {
         let startTunnelOptions = StartTunnelOptions(options: options ?? [:])
+
+        #if os(macOS)
         if !startTunnelOptions.isStartedByApp {
             if let tunnelProtocol = protocolConfiguration as? NETunnelProviderProtocol {
                 if tunnelProtocol.shouldPreventAutomaticConnections {
@@ -38,10 +42,21 @@ class PacketTunnelProvider: OpenVPNTunnelProvider {
                 }
             }
         }
+        #endif
 
-        super.startTunnel(options: options, completionHandler: completionHandler)
+        var appVersionString = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String ?? "Unknown version"
+        if let appBuild = Bundle.main.infoDictionary?["CFBundleVersion"] as? String {
+            appVersionString += " (\(appBuild))"
+        }
+        appVersion = appVersionString
+
+        super.startTunnel(options: options) { [weak self] error in
+            if startTunnelOptions.isStartedByApp {
+                self?.rewriteLog(useDiskLog: error != nil)
+            }
+            completionHandler(error)
+        }
     }
-    #endif
 
     override func handleAppMessage(_ messageData: Data, completionHandler: ((Data?) -> Void)? = nil) {
         // Convert TunnelKit's response to our response
@@ -78,5 +93,82 @@ class PacketTunnelProvider: OpenVPNTunnelProvider {
                 OpenVPNTunnelProvider.Message.requestLog.data,
                 completionHandler: completionHandler)
         }
+    }
+}
+
+private extension PacketTunnelProvider {
+    func rewriteLog(useDiskLog: Bool) {
+        if useDiskLog {
+            moveUpLastLogSeparatorInDiskLog()
+        } else {
+            moveUpLastLogSeparatorInMemoryLog()
+        }
+    }
+
+    func moveUpLastLogSeparatorInDiskLog() {
+        // If there was an error during startTunnel, the log would've been already written to disk.
+        // In that case, we should read the log on disk and modify that.
+        let debugLogFilename = "debug.log"
+        guard let tunnelProtocol = protocolConfiguration as? NETunnelProviderProtocol,
+           let providerConfiguration = tunnelProtocol.providerConfiguration,
+           let appGroup = try? Configuration.appGroup(from: providerConfiguration),
+           let parentURL = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: appGroup) else {
+               return
+        }
+        let debugLogURL = parentURL.appendingPathComponent(debugLogFilename)
+
+        if let logString = try? String(contentsOf: debugLogURL) {
+            var logLines = logString.components(separatedBy: "\n")
+            if moveUpLastLogSeparator(in: &logLines) {
+                let content = logLines.joined(separator: "\n")
+                try? content.write(to: debugLogURL, atomically: true, encoding: .utf8)
+
+            }
+        }
+    }
+
+    func moveUpLastLogSeparatorInMemoryLog() {
+        // In case there was no error during startTunnel, the log would still be in memory.
+        // So we modify the in-memory log.
+        let memoryLog = SwiftyBeaver.self.destinations.compactMap({ $0 as? MemoryDestination }).first
+        if let memoryLog = memoryLog {
+            let mirror = Mirror(reflecting: memoryLog)
+            for child in mirror.children {
+                // Read the private member called 'buffer' in the MemoryDestination instance
+                if child.label == "buffer",
+                   let buffer = child.value as? [String] {
+                    var lines = buffer
+                    if moveUpLastLogSeparator(in: &lines) {
+                        memoryLog.start(with: lines)
+                    }
+                }
+            }
+        }
+    }
+
+    func indexOfTrailingAppLog(in lines: [String], logSeparatorIndex: Int, appSeparator: String, otherSeparators: [String]) -> Int? {
+        for index in stride(from: logSeparatorIndex - 1, through: 0, by: -1) {
+            let line = lines[index]
+            if line == appSeparator {
+                return index
+            }
+            if otherSeparators.contains(line) {
+                return nil
+            }
+        }
+        return nil
+    }
+
+    func moveUpLastLogSeparator(in lines: inout [String]) -> Bool {
+        // Move the last logSeparator line above the last "App:" line, so that the
+        // app log goes together with the corresponding tunnel log.
+        if let lastLogSeparatorIndex = lines.lastIndex(of: logSeparator),
+            let appLogStartIndex = indexOfTrailingAppLog(in: lines, logSeparatorIndex: lastLogSeparatorIndex,
+                                                            appSeparator: "App:", otherSeparators: ["Tunnel:", logSeparator]) {
+            lines.replaceSubrange(lastLogSeparatorIndex ..< lastLogSeparatorIndex + 2, with: ["Tunnel:"])
+            lines.insert(contentsOf: [logSeparator, ""], at: appLogStartIndex)
+            return true
+        }
+        return false
     }
 }